requirements.txt
annotated-types==0.7.0
anyio==4.7.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.3
pydantic-core==2.27.1
python-dotenv==1.0.1
replicate==1.0.4
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
virtuals-sdk==0.1.3
